+++ src/lib/filter/button.svelte
@@ -1,0 +1,13 @@
+<script lang="ts">
+    import { filterStore } from "./store";
+    function toggle(): void
+    {
+        filterStore.update(filter => {
+            filter.visible = !filter.visible;
+            return filter;
+        });
+    }
+</script>
+<button type="button" on:click={toggle}>
+    {$filterStore.visible ? 'Ocultar' : 'Exibir'} filtros
+</button>